requirements.txt
beautifulsoup4
datasets
langchain
langchain-community
langchain-openai
langchain-pinecone
pinecone-client
openai
requests
tiktoken
tqdm
unidecode
jq
faiss-cpu
